requirements.txt
streamlit==1.35.0
langchain==0.2.0
langchain-community==0.2.0
langchain-groq==0.1.0
pypdf==4.2.0
chromadb==0.5.0
scikit-learn==1.5.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
huggingface_hub==0.23.